Fix large file streaming; fix integer overflow; DLNA support
Bug with large file streaming fixed in master branch of NanoHTTPD.
Since release 2.3.2 has not yet been released, temporarily source code
will be included directly in the libretorrent source code tree.

+++ app/src/main/java/org/proninyaroslav/libretorrent/core/TorrentInputStream.java
@@ -104,7 +104,7 @@ public class TorrentInputStream extends InputStream
         int firstPieceSize = (stream.firstFilePiece == stream.lastFilePiece ?
                               stream.lastFilePieceSize :
                               stream.pieceLength);
-        long firstPieceEnd = stream.firstFilePiece * stream.pieceLength + firstPieceSize;
+        long firstPieceEnd = (long)stream.firstFilePiece * stream.pieceLength + firstPieceSize;
 
         if (stream.fileOffset > firstPieceEnd)
             throw new IllegalArgumentException();
@@ -169,7 +169,7 @@ public class TorrentInputStream extends InputStream
         int pieceSize = (piece == stream.lastFilePiece ?
                          stream.lastFilePieceSize :
                          stream.pieceLength);
-        long pieceEnd = pieceLocalIndex * stream.pieceLength + pieceSize;
+        long pieceEnd = (long)pieceLocalIndex * stream.pieceLength + pieceSize;
 
         return pieceSize - (int)(pieceEnd - pos);
     }
